requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alembic==1.14.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docutils==0.21.2
ecdsa==0.19.0
fastapi==0.115.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
htmx==0.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
ifaddr==0.2.0
inflect==7.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
markdown2==2.5.1
MarkupSafe==3.0.2
more-itertools==10.5.0
multidict==6.1.0
nicegui==2.5.0
orjson==3.10.11
passlib==1.7.4
propcache==0.2.0
pscript==0.7.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==1.10.19
Pygments==2.18.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-jose==3.3.0
python-multipart==0.0.17
python-socketio==5.11.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==75.5.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.54
starlette==0.41.2
typeguard==4.4.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
vbuild==0.8.2
watchfiles==0.24.0
websockets==14.1
wsproto==1.2.0
yarl==1.17.1
